Fix for issue #99 with scaling for pictures. Changed datatype for internal pixel variables from int to double.

+++ src/EPPlus/Drawing/ExcelDrawings.cs
@@ -1060,7 +1060,7 @@ namespace OfficeOpenXml.Drawing
             }
         }
         #endregion 
-        internal void AdjustWidth(int[,] pos)
+        internal void AdjustWidth(double[,] pos)
         {
             var ix = 0;
             //Now set the size for all drawings depending on the editAs property.
@@ -1078,7 +1078,7 @@ namespace OfficeOpenXml.Drawing
                 ix++;
             }
         }
-        internal void AdjustHeight(int[,] pos)
+        internal void AdjustHeight(double[,] pos)
         {
             var ix = 0;
             //Now set the size for all drawings depending on the editAs property.
@@ -1096,9 +1096,9 @@ namespace OfficeOpenXml.Drawing
                 ix++;
             }
         }
-        internal int[,] GetDrawingWidths()
+        internal double[,] GetDrawingWidths()
         {
-            int[,] pos = new int[Count, 2];
+            double[,] pos = new double[Count, 2];
             int ix = 0;
             //Save the size for all drawings
             foreach (ExcelDrawing d in this)
@@ -1108,9 +1108,9 @@ namespace OfficeOpenXml.Drawing
             }
             return pos;
         }
-        internal int[,] GetDrawingHeight()
+        internal double[,] GetDrawingHeight()
         {
-            int[,] pos = new int[Count, 2];
+            double[,] pos = new double[Count, 2];
             int ix = 0;
             //Save the size for all drawings
             foreach (ExcelDrawing d in this)
